requirements.txt
google-genai>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
streamlit>=1.37.0
pydantic>=2.8.0

